+++ weave-js/src/util/refs.ts
@@ -1,4 +1,4 @@
-import {WEAVE_REF_PREFIX} from '../components/PagePanelComponents/Home/Browse3/pages/wfReactInterface/constants';
+import {WEAVE_PRIVATE_PREFIX, WEAVE_REF_PREFIX} from '../components/PagePanelComponents/Home/Browse3/pages/wfReactInterface/constants';
 
 const encodeSelect = (part: string): string => {
   const symbols: string[] = ['%', '/', ':'];
@@ -39,3 +39,16 @@ export const makeRefObject = (
 export const abbreviateRef = (ref: string): string => {
   return WEAVE_REF_PREFIX + '/...' + ref.slice(-6);
 };
+
+export const privateRefToSimpleName = (ref: string) => {
+  if (!ref.startsWith(WEAVE_PRIVATE_PREFIX)) {
+     throw new Error('Not a private ref');
+  }
+  const trimmed = ref.replace(`${WEAVE_PRIVATE_PREFIX}//`, '');
+  try {
+    return decodeURIComponent(trimmed.split('/')[1].split(':')[0]);
+  } catch (e) {
+    return trimmed;
+  }
+};
+
